Moved files to standard structure
