+++ OneOneOne/Pipfile
@@ -1,0 +1,15 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+
+[packages]
+integration-adaptors-common = {editable = true,path = "./../common"}
+
+[requires]
+python_version = "3.7"
+
+[scripts]
+tests = "python -m unittest -v"
