+++ OneOneOne/definitions.py
@@ -1,0 +1,5 @@
+import os
+from pathlib import Path
+
+ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
+XML_PATH = Path(ROOT_DIR) / 'response_xmls'
